Update ClutterAlpha
ClutterAlpha now uses a floating point value from -1.0 to 2.0 instead of
an integer between 0 and CLUTTER_ALPHA_MAX_ALPHA. The old clutter_*_func
alpha functions are also replaced by easing modes.

+++ clutter/pyclutter.h
@@ -23,7 +23,7 @@ void               pyclutter_callback_free   (PyClutterCallback *cb);
 PyObject *         pyclutter_callback_invoke (PyClutterCallback *cb,
                                               ...);
 
-guint32   pyclutter_alpha_func              (ClutterAlpha    *alpha,
+gdouble   pyclutter_alpha_func              (ClutterAlpha    *alpha,
                                              gpointer         data);
 
 gboolean  pyclutter_geometry_from_pyobject  (PyObject        *object,
